lab_8/service2/requirements.txt
psycopg2==2.9.1
kafka-python==2.0.2
